Refactor: factor out array_size.rs
